Change integer type to int64

+++ 10-08-2021/even_odd_numbers.pas
@@ -1,4 +1,4 @@
-var n, even, odd: integer;
+var n, even, odd: int64;
 
 function is_even(a: integer): boolean;
 begin
